chore: rename files, improve types and samples
