--- packages/lab/src/Canvas/CanvasPanel/index.tsx
@@ -1,1 +1,0 @@
-export * from "./CanvasPanel";
