--- packages/lab/src/Canvas/CanvasTabs/index.tsx
@@ -1,1 +1,0 @@
-export * from "./CanvasTabs";
