--- packages/lab/src/Canvas/CanvasToolbar/index.tsx
@@ -1,1 +1,0 @@
-export * from "./CanvasToolbar";
