+++ packages/lab/src/Canvas/Panel/Panel.styles.tsx
@@ -7,7 +7,7 @@ export const { staticClasses, useClasses } = createClasses("HvCanvasPanel", {
     position: "absolute",
     top: 0,
     left: 0,
-    boxShadow: `4px 0px 8px -4px color-mix(in srgb, #414141 12%, transparent)`,
+    boxShadow: "4px 0px 8px -4px color-mix(in srgb, #414141 12%, transparent)",
     zIndex: theme.zIndices.overlay,
     backgroundColor: "transparent",
     transition: "width 0.3s ease",
@@ -31,12 +31,12 @@ export const { staticClasses, useClasses } = createClasses("HvCanvasPanel", {
     display: "flex",
     justifyContent: "center",
     zIndex: theme.zIndices.overlay,
-    boxShadow: `4px 0px 8px -4px color-mix(in srgb, #414141 12%, transparent)`,
+    boxShadow: "4px 0px 8px -4px color-mix(in srgb, #414141 12%, transparent)",
     backgroundColor: theme.colors.atmo1,
-    borderRadius: `0 ${theme.radii.round} ${theme.radii.round} 0 `,
+    borderRadius: `0 ${theme.radii.round} ${theme.radii.round} 0`,
     position: "absolute",
     transition: "left 0.3s ease",
-    top: `calc(50% - 86px)`,
+    top: "calc(50% - 86px)",
     "&$handleOpen": {
       left: 320,
     },
@@ -48,7 +48,7 @@ export const { staticClasses, useClasses } = createClasses("HvCanvasPanel", {
     },
   },
   handleButton: {
-    top: `calc(50% - 16px)`,
+    top: "calc(50% - 16px)",
     position: "absolute",
   },
   open: {},
