+++ packages/lab/src/Canvas/Panel/Panel.tsx
@@ -9,8 +9,8 @@ import {
 } from "@hitachivantara/uikit-react-core";
 import { End, Start } from "@hitachivantara/uikit-react-icons";
 
-import { HvCanvasTab, HvCanvasTabs } from "../CanvasTabs";
-import { staticClasses, useClasses } from "./CanvasPanel.styles";
+import { HvCanvasTab, HvCanvasTabs, HvCanvasTabsProps } from "../Tabs";
+import { staticClasses, useClasses } from "./Panel.styles";
 
 export { staticClasses as canvasPanelClasses };
 
@@ -22,22 +22,22 @@ const DEFAULT_LABELS = {
 };
 
 export interface HvCanvasPanelProps extends HvBaseProps<HTMLDivElement> {
-  /** Whether the panel is open or not. If this property is defined the panel must be fully controlled. */
+  /** Whether the panel is open or not. If this property is defined, the panel must be fully controlled. */
   open?: boolean;
   /** When uncontrolled, defines the initial state of the panel. */
   defaultOpened?: boolean;
-  /** The tabs that should be visible on the canvas panel */
+  /** The tabs that should be visible on the canvas panel. */
   tabs?: HvCanvasTab[];
-  /** The function that will be executed whenever the panel toggles. It will receive the state of the accordion. */
+  /** The function that will be executed whenever the panel toggles. It will receive the state of the panel. */
   onToggle?: (
     event: React.MouseEvent | React.KeyboardEvent,
     open: boolean,
   ) => void;
-  /** The function that will be executed when a tab changes.It will receive the id of the selected tab. */
-  onTabChange?: (event: React.SyntheticEvent, tabId: string) => void;
+  /** The function that will be executed when a tab changes. It will receive the id of the selected tab. */
+  onTabChange?: (event: React.SyntheticEvent | null, tabId: string) => void;
   /** An object containing all the labels. */
   labels?: Partial<typeof DEFAULT_LABELS>;
-  /* The content that will be rendered within canvas panel. */
+  /* The content that will be rendered within the canvas panel. */
   children?: React.ReactNode;
   /** A Jss Object used to override or extend the styles applied. */
   classes?: HvCanvasPanelClasses;
@@ -83,12 +83,12 @@ export const HvCanvasPanel = (props: HvCanvasPanelProps) => {
     setPanelWidth();
   }, [open, setPanelWidth]);
 
-  const handleTogglePanel = (event) => {
+  const handleTogglePanel = (event: React.MouseEvent | React.KeyboardEvent) => {
     setOpen((prev) => !prev);
     onToggle?.(event, !open);
   };
 
-  const handleTabChange = (event, tabId) => {
+  const handleTabChange: HvCanvasTabsProps["onChange"] = (event, tabId) => {
     onTabChange?.(event, tabId);
   };
