+++ packages/lab/src/Canvas/Panel/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Panel";
