+++ packages/lab/src/Canvas/Toolbar/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Toolbar";
