+++ packages/lab/src/Canvas/stories/Canvas.stories.tsx
@@ -1,37 +1,8 @@
-import { useState } from "react";
-import { css } from "@emotion/css";
 import { Meta, StoryObj } from "@storybook/react";
 import { userEvent, within } from "@storybook/test";
-import { BackgroundVariant } from "reactflow";
-import {
-  HvButton,
-  HvDropDownMenu,
-  HvIconButton,
-  HvInlineEditor,
-  theme,
-} from "@hitachivantara/uikit-react-core";
-import {
-  Backwards,
-  Calendar,
-  DataSource,
-  Plane,
-  Pod,
-  Redo,
-  Schema,
-  Undo,
-  User,
-} from "@hitachivantara/uikit-react-icons";
-import {
-  HvCanvasPanel,
-  HvCanvasToolbar,
-  HvFlow,
-  HvFlowBackground,
-  HvFlowMinimap,
-} from "@hitachivantara/uikit-react-lab";
 
-import { ListView } from "./ListView/ListView";
-import { eds, nds, nodeTypes } from "./nodes";
-import { TreeView } from "./TreeView";
+import { MainStory } from "./Main";
+import MainRaw from "./Main?raw";
 
 const meta: Meta = {
   title: "Lab/Canvas",
@@ -42,33 +13,26 @@ const meta: Meta = {
 };
 export default meta;
 
-const tabs = [
-  {
-    id: "1",
-    content: (
-      <>
-        <DataSource />
-        Add Data
-      </>
-    ),
-  },
-  {
-    id: "2",
-    content: (
-      <>
-        <Schema />
-        Model Structure
-      </>
-    ),
-  },
-];
-
 export const Main: StoryObj = {
-  args: {
-    title: "Semantic Model Name",
-  },
-  argTypes: {
-    classes: { control: { disable: true } },
+  decorators: [
+    (Story) => (
+      <div
+        style={{
+          width: "100%",
+          height: "100vh",
+          position: "relative",
+        }}
+      >
+        {Story()}
+      </div>
+    ),
+  ],
+  parameters: {
+    docs: {
+      source: {
+        code: MainRaw,
+      },
+    },
   },
   // For visual testing
   play: async ({ canvasElement }) => {
@@ -78,88 +42,5 @@ export const Main: StoryObj = {
     });
     await userEvent.click(button);
   },
-  render: () => {
-    const [selectedTab, setSelectedTab] = useState<string>("1");
-
-    const classes = {
-      separator: css({
-        height: 30,
-        width: 1,
-        backgroundColor: theme.colors.atmo4,
-        margin: `0 ${theme.space.xs}`,
-      }),
-    };
-    const backButton = (
-      <HvButton aria-label="Back" icon>
-        <Backwards />
-      </HvButton>
-    );
-
-    const title = (
-      <HvInlineEditor value="Semantic Model Name" variant="title4" />
-    );
-
-    const Separator = () => <div className={classes.separator} />;
-
-    const renderIcon = (Icon: React.ElementType) => () => <Icon />;
-
-    return (
-      <div
-        style={{
-          width: "100%",
-          height: "100vh",
-          position: "relative",
-        }}
-      >
-        <HvFlow
-          style={{ width: "100%", height: "100vh" }}
-          snapGrid={[1, 1]}
-          snapToGrid
-          nodes={nds}
-          edges={eds}
-          nodeTypes={nodeTypes}
-        >
-          <HvFlowBackground
-            color={theme.colors.secondary_80}
-            variant={BackgroundVariant.Dots}
-            gap={16}
-            style={{ color: "red" }}
-          />
-          <HvFlowMinimap />
-        </HvFlow>
-        <HvCanvasToolbar backButton={backButton} title={title}>
-          <HvIconButton title="Undo">
-            <Undo />
-          </HvIconButton>
-          <HvIconButton title="Redo">
-            <Redo />
-          </HvIconButton>
-          <Separator />
-          <HvIconButton title="Add">
-            <Pod />
-          </HvIconButton>
-          <Separator />
-          <HvButton variant="primary">Save</HvButton>
-          <HvButton variant="primaryGhost">Cancel</HvButton>
-          <Separator />
-          <HvDropDownMenu
-            placement="right"
-            onClick={(e, item) => console.log(item.label)}
-            dataList={[
-              { label: "Label 1", icon: renderIcon(User) },
-              { label: "Label 2", icon: renderIcon(Calendar), disabled: true },
-              { label: "Label 3", icon: renderIcon(Plane) },
-            ]}
-          />
-        </HvCanvasToolbar>
-        <HvCanvasPanel
-          tabs={tabs}
-          onTabChange={(event, tabId) => setSelectedTab(tabId)}
-        >
-          {selectedTab === "1" && <TreeView />}
-          {selectedTab === "2" && <ListView />}
-        </HvCanvasPanel>
-      </div>
-    );
-  },
+  render: () => <MainStory />,
 };
